debug

+++ part1/utils.c
@@ -146,7 +146,7 @@ uint8_t numLRUBits(cache_t* cache) {
 uint64_t totalBlockBits(cache_t* cache) {
 	/* Your Code Here. */
     //printf("%d | ",getTagSize(cache));
-	return 8*(cache->blockDataSize)+3+numLRUBits(cache)+getTagSize(cache); // offset bits + dirty bit + valid bit +lru
+	return (uint64_t) (8*(cache->blockDataSize)+3+numLRUBits(cache)+getTagSize(cache)); // offset bits + dirty bit + valid bit +lru
 }
 
 /*
@@ -155,7 +155,7 @@ uint64_t totalBlockBits(cache_t* cache) {
 uint64_t cacheSizeBits(cache_t* cache) {
 	/* Your Code Here. */
     
-	return (cache->totalDataSize/ cache->blockDataSize)*totalBlockBits(cache) ;
+	return (uint64_t) ((cache->totalDataSize/ cache->blockDataSize)*totalBlockBits(cache)) ;
 }
 
 /*
